モジュールの分割(Page系と定数)
